+++ svm/src/conformance/mod.rs
@@ -6,6 +6,8 @@ pub mod callback;
 #[cfg(feature = "conformance")]
 pub mod elf_loader;
 #[cfg(feature = "conformance")]
+pub mod fd_hash;
+#[cfg(feature = "conformance")]
 pub mod feature_set;
 pub mod instr;
 pub mod programs;
